Check Function added
Check if the clicked video is already in the list but on second
thought, this function is redundant since the viewer might want to
watch the same video multiple times. Will have to delete it later.

Figured out the way to move on to the next queued video when the
current video is over. But still need to find out why the video-id in
the playlist is ‘undefined’